pyproject.toml
[tool.poetry]
name = "rivian-python-client"
version = "2.1.0"
description = "Rivian API Client (Unofficial)"
readme = "README.md"
authors = ["Brian Retterer <bretterer@gmail.com>"]
license = "MIT"
packages = [
    { include = "rivian", from = "src" },
]

[tool.poetry.dependencies]
python = ">=3.9,<3.14"
# Pin to versions compatible with Home Assistant 2025.10.4
aiohttp = ">=3.0.0,<=3.12.15"
attrs = ">=20.3.0,<=25.3.0"
propcache = ">=0.1.0,<=0.3.2"
yarl = ">=1.6.0,<=1.20.1"
cryptography = ">=41.0.1,<=45.0.7"
protobuf = ">=3.20.0,<=6.32.0"
backports-strenum = { version = "^1.2.4", python = "<3.11" }
bleak = { version = ">=0.21,<=1.0.1", optional = true }
dbus-fast = {version = "^2.11.0", optional = true, platform = "linux"}
gql = {version = ">=3.0.0,<=3.5.3", extras = ["aiohttp"]}
# Install websockets separately for HA compatibility
# gql[websockets] extra requires websockets<12, but HA integrations need >=13
# We install websockets separately and gql will use it for WebSocket subscriptions
websockets = ">=13.0,<16.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.1.2,<9.0.0"
pytest-asyncio = ">=0.18.3,<1.1.0"
python-dotenv = ">=0.20,<1.2"
aresponses = ">=2.1.5,<4.0.0"
mypy = ">=1.7.0,<2.0.0"
ruff = ">=0.4,<0.13"
pre-commit = "^4.1.0"
grpcio-tools = "^1.76.0"

[tool.poetry.extras]
ble = ["bleak", "dbus-fast"]

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
style = "semver"
pattern = "default-unprefixed"

[tool.ruff.lint.isort]
combine-as-imports = true
split-on-trailing-comma = false

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend"
